simplified directory structure
