+++ package.json
@@ -10,7 +10,7 @@
     "@feathersjs/express": "^1.2.2",
     "@feathersjs/feathers": "^3.1.5",
     "debug": "^3.1.0",
-    "deified": "^0.3.0",
+    "deified": "^0.3.1",
     "micromatch": "^3.1.10",
     "uuid": "^2.0.1"
   },
@@ -18,10 +18,10 @@
     "coveralls": "^3.0.1",
     "eslint": "^3.14.1",
     "eslint-config-airbnb-base": "^11.0.1",
-    "eslint-plugin-import": "^2.2.0",
-    "jest": "^22.4.3",
+    "eslint-plugin-import": "^2.12.0",
+    "jest": "^22.4.4",
     "jsdoc": "^3.3.2",
-    "node-mocks-http": "^1.5.0"
+    "node-mocks-http": "^1.7.0"
   },
   "scripts": {
     "start": "node test/express.js",
