+++ test/api/alpha.solos.js
