+++ test/api/beta.solos.js
